refactor code to be more reusable
