+++ lib/store_app/common/app_page/app_loading_page.dart
@@ -18,71 +18,14 @@
 import 'package:flutter/material.dart';
 import 'package:shimmer/shimmer.dart';
 import 'package:software/store_app/common/app_page/page_layouts.dart';
+import 'package:software/store_app/common/app_page/app_swipe_gesture.dart';
 import 'package:software/store_app/common/border_container.dart';
-import 'package:software/store_app/common/constants.dart';
 import 'package:yaru_icons/yaru_icons.dart';
 
-class AppLoadingPage extends StatefulWidget {
+class AppLoadingPage extends StatelessWidget {
   const AppLoadingPage({
     super.key,
   });
-
-  @override
-  State<AppLoadingPage> createState() => _AppLoadingPageState();
-}
-
-class _AppLoadingPageState extends State<AppLoadingPage> {
-  final double width = 50;
-  final double height = 50;
-
-  double xPosition = 0;
-  double yPosition = 0;
-
-  double currentExtent = 0;
-
-  bool isVisible = false;
-
-  void onPanUpdate(DragUpdateDetails details) {
-    if (details.delta.dx > 0 &&
-        details.delta.dy < 50 &&
-        details.delta.dy > -50 &&
-        currentExtent <= kMaxExtent) {
-      currentExtent += details.delta.dx;
-      setState(() {
-        xPosition += details.delta.dx * 0.2;
-      });
-    }
-    if (details.delta.dx < 0 &&
-        details.delta.dy < 50 &&
-        details.delta.dy > -50 &&
-        currentExtent >= -width) {
-      currentExtent -= -details.delta.dx;
-      setState(() {
-        xPosition -= -details.delta.dx * 0.2;
-      });
-    }
-  }
-
-  void onPanStart(DragStartDetails details) {
-    currentExtent = 0;
-    xPosition = 0 - width;
-    yPosition =
-        (MediaQueryData.fromWindow(WidgetsBinding.instance.window).size.height -
-                height) /
-            2;
-    setState(() {
-      isVisible = true;
-    });
-  }
-
-  void onPanEnd(DragEndDetails details) {
-    if (currentExtent > (kMaxExtent / 2)) {
-      Navigator.of(context).pop();
-    }
-    setState(() {
-      isVisible = false;
-    });
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -211,33 +154,8 @@ class _AppLoadingPageState extends State<AppLoadingPage> {
           onPressed: () => Navigator.pop(context),
         ),
       ),
-      body: GestureDetector(
-        onPanUpdate: onPanUpdate,
-        onPanStart: onPanStart,
-        onPanEnd: onPanEnd,
-        child: Stack(
-          children: <Widget>[
-            body,
-            Positioned(
-              top: yPosition,
-              left: xPosition,
-              child: Visibility(
-                visible: isVisible,
-                child: Container(
-                  decoration: const BoxDecoration(
-                    color: Color.fromRGBO(255, 127, 80, 1),
-                    shape: BoxShape.circle,
-                  ),
-                  width: 50,
-                  height: 50,
-                  child: const Icon(
-                    Icons.arrow_back_rounded,
-                  ),
-                ),
-              ),
-            ),
-          ],
-        ),
+      body: BackGesture(
+        body: body,
       ),
     );
   }
