+++ lib/store_app/common/app_page/app_page.dart
@@ -24,9 +24,9 @@ import 'package:software/store_app/common/app_page/app_infos.dart';
 import 'package:software/store_app/common/app_page/app_reviews.dart';
 import 'package:software/store_app/common/app_page/media_tile.dart';
 import 'package:software/store_app/common/app_page/page_layouts.dart';
+import 'package:software/store_app/common/app_page/app_swipe_gesture.dart';
 import 'package:software/store_app/common/border_container.dart';
 import 'package:software/store_app/common/safe_network_image.dart';
-import 'package:software/store_app/common/constants.dart';
 import 'package:yaru_icons/yaru_icons.dart';
 import 'package:yaru_widgets/yaru_widgets.dart';
 
@@ -79,58 +79,6 @@ class AppPage extends StatefulWidget {
 
 class _AppPageState extends State<AppPage> {
   late YaruCarouselController controller;
-
-  final double width = 50;
-  final double height = 50;
-
-  double xPosition = 0;
-  double yPosition = 0;
-
-  double currentExtent = 0;
-
-  bool isVisible = false;
-
-  void onPanUpdate(DragUpdateDetails details) {
-    if (details.delta.dx > 0 &&
-        details.delta.dy < 50 &&
-        details.delta.dy > -50 &&
-        currentExtent <= kMaxExtent) {
-      currentExtent += details.delta.dx;
-      setState(() {
-        xPosition += details.delta.dx * 0.2;
-      });
-    }
-    if (details.delta.dx < 0 &&
-        details.delta.dy < 50 &&
-        details.delta.dy > -50 &&
-        currentExtent >= -width) {
-      currentExtent -= -details.delta.dx;
-      setState(() {
-        xPosition -= -details.delta.dx * 0.2;
-      });
-    }
-  }
-
-  void onPanStart(DragStartDetails details) {
-    currentExtent = 0;
-    xPosition = 0 - width;
-    yPosition =
-        (MediaQueryData.fromWindow(WidgetsBinding.instance.window).size.height -
-                height) /
-            2;
-    setState(() {
-      isVisible = true;
-    });
-  }
-
-  void onPanEnd(DragEndDetails details) {
-    if (currentExtent > (kMaxExtent / 2)) {
-      Navigator.of(context).pop();
-    }
-    setState(() {
-      isVisible = false;
-    });
-  }
 
   @override
   void initState() {
@@ -284,42 +232,16 @@ class _AppPageState extends State<AppPage> {
             : narrowWindowLayout;
 
     return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.appData.title),
-        titleSpacing: 0,
-        leading: _CustomBackButton(
-          onPressed: () => Navigator.pop(context),
+        appBar: AppBar(
+          title: Text(widget.appData.title),
+          titleSpacing: 0,
+          leading: _CustomBackButton(
+            onPressed: () => Navigator.pop(context),
+          ),
         ),
-      ),
-      body: GestureDetector(
-        onPanUpdate: onPanUpdate,
-        onPanStart: onPanStart,
-        onPanEnd: onPanEnd,
-        child: Stack(
-          children: <Widget>[
-            body,
-            Positioned(
-              top: yPosition,
-              left: xPosition,
-              child: Visibility(
-                visible: isVisible,
-                child: Container(
-                  decoration: const BoxDecoration(
-                    color: Color.fromRGBO(255, 127, 80, 1),
-                    shape: BoxShape.circle,
-                  ),
-                  width: 50,
-                  height: 50,
-                  child: const Icon(
-                    Icons.arrow_back_rounded,
-                  ),
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
+        body: BackGesture(
+          body: body,
+        ));
   }
 }
